+++ init.php
@@ -1,0 +1,4 @@
+<?php namespace forall\events;
+
+//Register a special event dispatcher with the core.
+$core->register('dispatcher', new EventDispatcher);