+++ src/forall/events/EventDispatcher.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @package forall.events
+ * @author Avaq <aldwin.vlasblom@gmail.com>
+ */
+namespace forall\events;
+
+/**
+ * EventDispatcher class.
+ */
+class EventDispatcher implements EventDispatcherInterface
+{
+  
+  use EventDispatcherTraits;
+  
+}